in the middle of splitting navbar.js into two files.
