+++ src/components/index.js
@@ -1,3 +1,3 @@
 export { DashHeader as Header } from './header/header';
-export { Navbar } from './navbar/navbar';
+export { NavbarContainer } from './navbar/navbar-container';
 export { RightSidebar } from './right-sidebar/right-sidebar';
