--- src/components/toastr/style/toastr.scss
@@ -1,7 +1,0 @@
-.toastr {
-  border-radius: 10px !important;
-}
-
-.extra-padding {
-  padding: 4px 4px 0 0 !important;
-}
